plugin: Fix LFO not working when optimization -O3 enabled
LFO only works in Debug build, but failed to work in Release. It's
strange. I've done these checks:

- set_freq(): Checked get_dco_out(3), dco_out_position[3],
  dco_out_length[3].
  - In Debug: those values changes properly.
  - In Release: dco_out_position[3] always remains 0, causing
    get_dco_out(3) always returns the first sample.
- Wavetables: All wavetables can be loaded with no errors.

So where on earth does the problem resides? Answer is the variable
`delay` in MinatonPlugin::_processAudioFrame().

`delay` is used to calculate the delay time before every LFO cycle. It
has never accessed by other functions or variables, just simply adding
to itself, or reset when it reaches 20.

Such a behavior is most likely a target of optimization when building
with GCC in -O3. At this time, `delay` will always be 0, so LFO cycles
will never run.

However, on Debug, `delay` is never optimized, and it works as-is.

To be concluded, it's easy to explain why LFO works in Debug, but
on contrary ignored in Release.

A way to stop this kind of unexpected optimization is to define `delay`
with keyword "volatile". It tolds compilers not to discard variables
even though it "seems to be unused".

After applying "volatile", everything works well.