widget: menu: Deal with long menus
Sometimes, menu can be extremely long. So we need to check how to move the menu,
to make sure that the whole menu can be shown. To be brief, we need to check
the room between mouse boundary top and cursor.

The original implementation does not consider this (Wolf Shaper only use
a brief menu, but I need to implement a long Preset menu). So I must do
this improvements.