plugin: Fully resample audio buffer, then output on demand
The former resample method is easy, but had a severe issue: you will
encounter stitches on high samplerates (>= 88200.0Hz).

This is because @cpuimage's resampler method will output an unrelated
value on the last output frame (usually half of the frame value before
it). Just this frame caused stitch, since the frames at the boundary of
two sample groups failed to become a perfect curve.

To solve this, the resampler procedure should be rewritten, to make sure
our resampled frames can be played completely.

Procedures (on every run() cycle):
1. Check if a resampled buffer is fully played or not. If not, send the
   remaining frames to host buffers until:
  - the buffer is full, or
  - all samples played.
2. If all resampled frames played, generate new samples, then resample
   it. Input sample counts is as same as host buffer size (equals param
   "frame").
3. At this time, if the host buffer is still not fully filled, redo the
   first step.

For testing, this commit also disabled AudioMidiSyncBuffer. Later I'll
test if my new method can work with it.

NOTE: This is a backup branch. Also included this change:
- plugin: Do not use realloc() to prevent "invalid next size" error